Finished preprocessing code Added filetomat object that turns a file into an OpenCV Mat object Added image object to represent a Mat object so that additional process can be called on the image Added a cross platform logger that logs to a log.txt file when compiled on Linux and uses the Android logger when compiled for Android Added default uncrusitfy config for prettifying cpp and header files Added a TakePhoto activity to the Android application that takes a photo Added a SavePhoto object that's called when TakePhoto wants to save the byte array containing the image data Added a README.txt file that describes how to build the project
Images are read from the files directory of the application, SavePhoto saves the byte array to this folder, a sample image called test.jpg is saved there when the application is first ran. This test image is temporary.
The full path to the image to be processed is sent to the native code through the JNI and then to the Image object.